+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Changelog
 
 - Fix create server button being available when it shouldn't
 - Log which pack/version is being launched when a user launches an instance
+- When a download fails and a progress bar is shown, add those bytes done to the total so not to overflow the progress